Single parse impl using less mutation.
Allows parsing of all top level values in a message.
This can be used to read schemas where the schema layout is embedded in
the code base, using id and types directly. Vs loading it from a
description file.
We do not yet support loading a schema from a description file.